feat: Update auth for account access & Add payday to account

+++ resolvers/bill.js
@@ -1,10 +1,11 @@
-import { checkAuth } from '../middleware/isAuth';
+import { checkAuth, checkAccountAccess } from '../middleware/isAuth';
 import { Account } from '../models/Account';
 import { OneOffPayment } from '../models/OneOffPayment';
 import { Bill } from '../models/Bill';
 
 const findBills = async (_, { accountId }, req) => {
-  checkAuth(req);
+  await checkAuth(req);
+  await checkAccountAccess(accountId, req);
 
   // Fetch the account associated with the authenticated user
   const userAccount = await Account.findOne({ id: accountId });
@@ -24,16 +25,19 @@ const findBills = async (_, { accountId }, req) => {
 };
 
 const findBill = async (_, { id }, req) => {
-  checkAuth(req);
+  await checkAuth(req);
   const bill = await Bill.findById(id);
   if (!bill) {
     throw new Error(`Bill with id '${id}' does not exist`);
   }
+  await checkAccountAccess(bill.account, req);
   return bill;
 };
 
 const createBill = async (_, { bill }, req) => {
-  checkAuth(req);
+  await checkAuth(req);
+  await checkAccountAccess(bill.account, req);
+
   try {
     const existingBill = await Bill.findOne({ name: bill.name, account: bill.account });
     if (existingBill) {
@@ -71,13 +75,14 @@ const createBill = async (_, { bill }, req) => {
 };
 
 const editBill = async (_, { id, bill }, req) => {
-  checkAuth(req);
-  try {
-    const currentBill = await Bill.findById(id);
-    if (!currentBill) {
-      throw new Error(`Bill with id '${id}' does not exist`);
-    }
+  await checkAuth(req);
+  const currentBill = await Bill.findById(id);
+  if (!currentBill) {
+    throw new Error(`Bill with id '${id}' does not exist`);
+  }
+  await checkAccountAccess(currentBill.account, req);
 
+  try {
     const mergedBill = Object.assign(currentBill, bill);
     mergedBill.__v = mergedBill.__v + 1;
 
@@ -99,13 +104,14 @@ const editBill = async (_, { id, bill }, req) => {
 };
 
 const deleteBill = async (_, { id }, req) => {
-  checkAuth(req);
-  try {
-    const bill = await Bill.findById(id);
-    if (!bill) {
-      throw new Error(`Bill with id '${id}' does not exist`);
-    }
+  await checkAuth(req);
+  const bill = await Bill.findById(id);
+  if (!bill) {
+    throw new Error(`Bill with id '${id}' does not exist`);
+  }
+  await checkAccountAccess(bill.account, req);
 
+  try {
     const response = await Bill.deleteOne({ _id: id });
     if (bill && response.deletedCount == 1) {
       return {
@@ -120,6 +126,68 @@ const deleteBill = async (_, { id }, req) => {
   }
 };
 
+const batchUpdateBills = async (_, { input }, req) => {
+  await checkAuth(req);
+  const { ids, paid } = input;
+
+  try {
+    // Verify all bills exist and belong to the user's account
+    const bills = await Bill.find({ _id: { $in: ids } });
+    if (bills.length !== ids.length) {
+      throw new Error('One or more bills not found');
+    }
+
+    // Check access for each bill
+    await Promise.all(bills.map(bill => checkAccountAccess(bill.account, req)));
+
+    // Update all bills
+    const result = await Bill.updateMany(
+      { _id: { $in: ids } },
+      { $set: { paid }, $inc: { __v: 1 } }
+    );
+
+    // Fetch updated bills
+    const updatedBills = await Bill.find({ _id: { $in: ids } });
+
+    return {
+      bills: updatedBills,
+      success: true,
+      updatedCount: result.modifiedCount
+    };
+  } catch (err) {
+    throw err;
+  }
+};
+
+const batchDeleteBills = async (_, { ids }, req) => {
+  await checkAuth(req);
+
+  try {
+    // Verify all bills exist and belong to the user's account
+    const bills = await Bill.find({ _id: { $in: ids } });
+    if (bills.length !== ids.length) {
+      throw new Error('One or more bills not found');
+    }
+
+    // Check access for each bill
+    await Promise.all(bills.map(bill => checkAccountAccess(bill.account, req)));
+
+    // Delete all bills
+    const result = await Bill.deleteMany({ _id: { $in: ids } });
+
+    // Update account's bills array
+    const accountIds = [...new Set(bills.map(bill => bill.account))];
+    await Account.updateMany({ _id: { $in: accountIds } }, { $pull: { bills: { $in: ids } } });
+
+    return {
+      success: true,
+      deletedCount: result.deletedCount
+    };
+  } catch (err) {
+    throw err;
+  }
+};
+
 exports.resolvers = {
   Query: {
     bills: findBills,
@@ -128,6 +196,8 @@ exports.resolvers = {
   Mutation: {
     createBill,
     editBill,
-    deleteBill
+    deleteBill,
+    batchUpdateBills,
+    batchDeleteBills
   }
 };
